+++ beab_sh/srcs/shell_export.c
@@ -6,7 +6,7 @@
 /*   By: ringii <ringii@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 14:02:31 by abonard           #+#    #+#             */
-/*   Updated: 2022/12/09 14:59:03 by ringii           ###   ########.fr       */
+/*   Updated: 2022/12/13 11:20:38 by ringii           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	ft_export_var(t_token *t, t_env *env, bool is_forked)
 {
-	char		*namevar;
-	char		*value;
-	size_t		i;
-	size_t		j;
+	char	*namevar;
+	char	*value;
+	int		i;
+	int		j;
 
 	i = 1;
 	while (i < t->cmd_ac)
